Remove default value for g:tumblr_http_proxy
The python script checks whether this value is defined, rather than
checking whether it's empty.  Having a default value means that this
check will always pass, and it will try to use an empty string as its
proxy.

This commit removes the default value so that python can correctly
determine that nothing has been set.