Reformatted documentation to match vim style
I was initially just going to put in vim help keywords to make it easier
to type ":help Tumble" and so forth, but while I was at it I thought I
might as well make the documentation match other documentation in vim.

I modelled the formatting on the official vim documentation (two tab
indent for command descriptions, etc).  I've added help keywords where
it seemed pertinent.

As far as possible I've tried to leave the wording as it was, though
there are some places where it felt natural to move parts of paragraphs
to later subsections of the documentation now that there were headings
in place there.